+++ src/IDBRequest.js
@@ -1,4 +1,5 @@
 import * as util from './util.js';
+import EventTarget from 'eventtarget';
 
 /**
  * The IDBRequest Object that is returns for all async calls
@@ -10,6 +11,7 @@ class IDBRequest {
         this.__result = undefined;
         this.__error = this.__source = this.__transaction = null;
         this.__readyState = 'pending';
+        this.setOptions({extraProperties: ['debug']}); // Ensure EventTarget preserves our properties
     }
     toString () {
         return '[object IDBRequest]';
@@ -18,12 +20,15 @@ class IDBRequest {
 
 util.defineReadonlyProperties(IDBRequest.prototype, ['result', 'error', 'source', 'transaction', 'readyState']);
 
+Object.assign(IDBRequest.prototype, EventTarget.prototype);
+
 /**
  * The IDBOpenDBRequest called when a database is opened
  */
 class IDBOpenDBRequest extends IDBRequest {
     constructor () {
         super();
+        this.setOptions({extraProperties: ['oldVersion', 'newVersion', 'debug']}); // Ensure EventTarget preserves our properties
         this.onblocked = this.onupgradeneeded = null;
     }
     toString () {