+++ tests-mocha/IDBFactory/deleteDatabase-spec.js
@@ -46,7 +46,7 @@ describe('IDBFactory.deleteDatabase', function () {
 
             del.onsuccess = function (event) {
                 if (env.isShimmed || (!env.browser.isIE && !env.browser.isSafari)) {
-                    expect(event).to.be.an.instanceOf(IDBVersionChangeEvent);   // IE and Safari use a normal event
+                    expect(event).to.be.an.instanceOf(indexedDB.modules.ProxyPolyfill);   // IE and Safari use a normal event
                 }
                 done();
             };